Avoid referencing named behaviors via window.farmOS.map
**Why?** Allow farmOS-map to be imported into bundled applications
and entirely avoid the use of window.farmOS.map for that case.
